Building a large cluster of nodes.
Collaborative MOOCs require a flexible network to support the adhoc
(churn) nature of collaborative learning. This backbone provides the
fastest possible communication within a
changing, unbounded, network topology. the ability for clusters of
groups to join (discovery)  other groups and participate in cooperative
problem solving. This problem solving and resolution is a feature that
is found in many cooperative-based systems where scheduling of
resources, distributing data, or ensemble-based solving is desired. We
describe problem solving being composed of several steps. They are: any
cluster can propose problems,
interact with other clusters (a.k.a. discussions), and cooperatively
reach a solution (consensus gathering).